compile & debug
